refactor(InputNumber): 입력 값을 처리하는 InputNumberParser 추가 및 입력 값을 담을 일급 컬렉션 InputNumber 객체 추가
입력 값을 처리하는 InputNumberParser를 추가하였습니다. 이 클래스는 String 형태의 inputNumber를 받아 List<Integer>로 변환합니다.
또한, 책임 분리에 따라, 입력 값 검증 로직을 NumberScanner에서 InputNumberParser로 이동하였습니다.
