+++ src/main/java/baseball/Application.java
@@ -2,15 +2,15 @@ package baseball;
 
 import baseball.count.ball.BallCounter;
 import baseball.count.strike.StrikeCounter;
-import baseball.input.InputValueParser;
-import baseball.input.NumberScanner;
+import baseball.number.inputnumber.InputNumberParser;
+import baseball.scanner.NumberScanner;
 import baseball.print.MessagePrinter;
 
 public class Application {
     public static void main(String[] args) {
         BaseballGame baseballGame = new BaseballGame(
                 new NumberScanner(),
-                new InputValueParser(),
+                new InputNumberParser(),
                 new BallCounter(),
                 new StrikeCounter(),
                 new MessagePrinter()
