+++ src/main/java/baseball/number/inputnumber/InputNumber.java
@@ -1,0 +1,17 @@
+package baseball.number.inputnumber;
+
+import java.util.Collections;
+import java.util.List;
+
+public class InputNumber {
+
+    private final List<Integer> inputNumbers;
+
+    InputNumber(List<Integer> inputNumbers) {
+        this.inputNumbers = inputNumbers;
+    }
+
+    public List<Integer> inputNumbers() {
+        return Collections.unmodifiableList(inputNumbers);
+    }
+}
